refactor: combined component login-button with user-menu
